aider: Set the `propagate` attribute to `False` for the Flask app's logger and the `AirtableOperations` class's logger to prevent log message duplication.
# Aider chat conversation:

USER: in log got this? '
2023-12-20 18:59:02.015 [ttb] [2023-12-20 23:59:02,014] INFO in app: Processing order message for symbol: EURNZD.PRO
2023-12-20 18:59:02.015 [ttb] 2023-12-20 23:59:02 Processing order message for symbol: EURNZD.PRO
2023-12-20 18:59:02.593 [ttb] [2023-12-20 23:59:02,592] INFO in app: Sending order to PineConnector with bypassed filters for symbol: EURNZD.PRO
2023-12-20 18:59:02.593 [ttb] 2023-12-20 23:59:02 Sending order to PineConnector with bypassed filters for symbol: EURNZD.PRO
2023-12-20 18:59:03.103 [ttb] 10.207.196.129 - - [20/Dec/2023 23:59:03] "POST /webhook HTTP/1.1" 200 -'

ASSISTANT: It appears that the log entries are being duplicated, which suggests that the logger might be configured to propagate messages to the root logger, which also has handlers attached. This can happen if the `propagate` attribute of the logger is set to `True`. To prevent this duplication, we can set the `propagate` attribute to `False` for the logger used in the `app.py` file.

Here is the diff to update the `app.py` file to prevent log message duplication:

```diff
--- app.py
+++ app.py
@@ ... @@
 app = Flask(__name__)
 handler = logging.StreamHandler()
 handler.setLevel(logging.INFO)  # Set to INFO to reduce the verbosity of the logs
 formatter = logging.Formatter('%(asctime)s %(message)s', datefmt='%Y-%m-%d %H:%M:%S')
 handler.setFormatter(formatter)
 app.logger.addHandler(handler)
 app.logger.setLevel(logging.INFO)  # Set to INFO to reduce the verbosity of the logs
-app.logger.propagate = True
+app.logger.propagate = False

 class AirtableOperations:
     def __init__(self):
         self.airtable = Airtable(config.AIRTABLE_BASE_ID, config.AIRTABLE_TABLE_NAME, api_key=config.AIRTABLE_API_KEY)
         self.logger = logging.getLogger(__name__)
+        self.logger.propagate = False
```

This change sets the `propagate` attribute to `False` for both the Flask app's logger and the `AirtableOperations` class's logger, which should prevent the log messages from being handled by the root logger's handlers and thus stop the duplication.